fix: fix router warnings about the require cycles
