+++ src/js/components/HeaderCartButton.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react';
 import { StyleSheet, View, Text, TouchableOpacity, Image } from 'react-native';
 import { ShoppingCart } from '../shopping-cart.js';
-import { IS_IOS, MUSEO_SANS_NORMAL } from '../config/Constants';
+import { IS_IOS, MUSEO_SANS_NORMAL, SCREENS } from '../config/Constants';
 import { testProperties } from '../config/TestProperties';
 import I18n from '../config/I18n';
 import { colors } from '../utils/colors';
-import { SCREENS } from '../Router';
 
 export default class CartButton extends Component {
   constructor(props) {
