+++ src/js/config/Constants.js
@@ -1,4 +1,4 @@
-import { Platform, Dimensions } from 'react-native';
+import { Dimensions, Platform } from 'react-native';
 
 export const TEST_PREFIX = 'test-';
 export const MAKE_ACCESSIBLE_FOR_AUTOMATION = { accessible: false };
@@ -17,3 +17,14 @@ export const IS_IPHONEX = IS_IOS &&
 // Check https://stackoverflow.com/questions/33088510/react-native-android-fontfamily-does-not-take-effect
 export const MUSEO_SANS_NORMAL = IS_IOS ? 'MuseoSans-500' : 'MuseoSans_500';
 export const MUSEO_SANS_BOLD = IS_IOS ? 'MuseoSans-700' : 'MuseoSans_700';
+export const SCREENS = {
+	LOGIN: 'Login',
+	INVENTORY_LIST: 'InventoryList',
+	INVENTORY_ITEM: 'InventoryItem',
+	CART_CONTENTS: 'CartContents',
+	CHECKOUT_SCREEN_ONE: 'CheckoutScreenOne',
+	CHECKOUT_SCREEN_TWO: 'CheckoutScreenTwo',
+	CHECKOUT_COMPLETE: 'CheckoutComplete',
+	WEBVIEW_SELECTION: 'WebviewSelection',
+	WEBVIEW_SCREEN: 'WebviewScreen',
+};
