+++ src/js/config/QuickActionsNavigation.js
@@ -1,0 +1,5 @@
+import { DeviceEventEmitter } from 'react-native';
+
+export function handleQuickActionsNavigation(navigation) {
+	DeviceEventEmitter.addListener('quickActionShortcut', data => navigation.replace(data.userInfo.url));
+}
