+++ src/js/screens/CheckoutScreenOne.js
@@ -11,8 +11,9 @@ import ArrowButton from '../components/ArrowButton';
 import ErrorMessageContainer from '../components/ErrorMessageContainer';
 import InputError from '../components/InputError';
 import SecondaryHeader from '../components/SecondaryHeader';
-import { handleQuickActionsNavigation, SCREENS } from '../Router';
 import { ShoppingCart } from '../shopping-cart';
+import { SCREENS } from '../config/Constants';
+import { handleQuickActionsNavigation } from '../config/QuickActionsNavigation';
 
 export default class CheckoutScreenOne extends Component {
   constructor(props) {
